+++ backend/canisters/community/impl/src/updates/cancel_p2p_swap.rs
@@ -20,7 +20,7 @@ fn cancel_p2p_swap(args: Args) -> Response {
 fn cancel_p2p_swap_impl(args: Args, state: &mut RuntimeState) -> OCResult<u32> {
     state.data.verify_not_frozen()?;
 
-    let member = state.get_calling_member(false)?;
+    let member = state.get_calling_member(None, false)?;
     let channel = state.data.channels.get_mut_or_err(&args.channel_id)?;
     let now = state.env.now();