+++ backend/canisters/community/impl/src/updates/register_webhook.rs
@@ -19,7 +19,7 @@ fn register_webhook(args: Args) -> Response {
 fn register_webhook_impl(args: Args, state: &mut RuntimeState) -> OCResult<SuccessResult> {
     state.data.verify_not_frozen()?;
 
-    let user_id = state.get_calling_member(true)?.user_id;
+    let user_id = state.get_calling_member(None, true)?.user_id;
     let channel = state.data.channels.get_mut_or_err(&args.channel_id)?;
     let member = channel.chat.members.get_verified_member(user_id)?;