+++ backend/canisters/community/impl/src/updates/update_community.rs
@@ -117,7 +117,7 @@ struct PrepareResult {
 fn prepare(args: &Args, state: &RuntimeState) -> OCResult<PrepareResult> {
     state.data.verify_not_frozen()?;
 
-    let member = state.get_calling_member(true)?;
+    let member = state.get_calling_member(None, true)?;
 
     let permissions = &state.data.permissions;
     if !member.role().can_update_details(permissions)