+++ backend/canisters/group/impl/src/updates/register_proposal_vote.rs
@@ -50,7 +50,7 @@ struct PrepareResult {
 fn prepare(args: &Args, state: &RuntimeState) -> OCResult<PrepareResult> {
     state.data.verify_not_frozen()?;
 
-    let member = state.get_calling_member(true)?;
+    let member = state.get_calling_member(None, true)?;
     let min_visible_event_index = member.min_visible_event_index();
 
     if let Some(proposal) = state