+++ backend/canisters/group/impl/src/updates/report_message.rs
@@ -40,7 +40,7 @@ async fn report_message_impl(args: Args) -> Response {
 fn build_c2c_args(args: &Args, state: &RuntimeState) -> OCResult<(c2c_report_message::Args, CanisterId)> {
     state.data.verify_not_frozen()?;
 
-    let member = state.get_calling_member(true)?;
+    let member = state.get_calling_member(None, true)?;
     let chat = &state.data.chat;
 
     if args.delete && !member.role().can_delete_messages(&chat.permissions) {